Moved enum to separate file
